[REF] Removed _ functions, moved resels, peak_clus
