+++ brainstat/stats/__init__.py
@@ -1,0 +1,1 @@
+""" The statistics tools of BrainStat """
