+++ brainstat/stats/multiple_comparisons.py
@@ -1,43 +1,52 @@
+"""Multiple comparison corrections."""
 import numpy as np
+from numpy import concatenate as cat
 from scipy.interpolate import interp1d
 from scipy.special import betaln, gammaln, gamma
+from scipy.linalg import toeplitz
+from scipy.sparse import csr_matrix
 import math
-from .models import _peak_clus, _resels
-from .utils import interp1, colon
+import copy
+from .utils import interp1, colon, ismember, row_ismember
+from ..mesh.utils import mesh_edges
 
-
-def _fdr(slm, mask=None):
+def fdr(slm, mask=None):
     """Q-values for False Discovey Rate of resels.
 
     Parameters
     ----------
-    slm : a dictionary with mandatory keys 't', 'df', and 'k'.
-        slm['t'] : numpy array of shape (1,v),
-            v is the number of vertices.
-        slm['df'] : numpy array of shape (1,1),
-            degrees of freedom.
-        slm['k'] : int,
-            number of variates.
-    Optional parameters:
-        mask : numpy array of shape (v), dtype 'bool',
-            by default ones(1,v).
-        slm['dfs'] : numpy array of shape (1,v),
-            effective degrees of freedom.
-        slm['resl'] : numpy array of shape (e,v),
-            matrix of sum over observations of squares of
-            differences of normalized residuals along each edge.
-        slm['tri'] : numpy array of shape (t,3),
-            triangle indices, 1-based, t is the number of triangles,
-        or,
-        slm['lat'] : 3D numpy array of 1's and 0's (1:in, 0:out).
+    slm : dict
+        Standard linear model returned by the t_test function; see Notes for
+        details.
+    mask : numpy.array, optional
+        Only t-values where the mask is true are considered. Defaults to a
+        vector of ones.
 
     Returns
     -------
-    qval : a dictionary with keys 'Q' and 'mask'
-        qval['Q'] : numpy array of shape (1,v),
-            vector of Q-values.
-        qval['mask'] : copy of mask.
+    qval : dict
+        Contains the Q-values in field 'Q' and a copy of the mask in 'mask'.
 
+    See Also
+    --------
+    brainstat.stats.models.t_test : Computes t-values for a linear model.
+
+    Notes
+    ------
+    The slm dictionary must contain at least the following fields:
+
+    - slm['t'] (numpy.array): a (1,v) array of t-values
+    - slm['df'] (numpy.array) of shape (1,1) containing the degrees of freedom
+    - slm['k'] (int) the number of variates.
+
+    Furthermore, slm may contain the following optional fields.
+
+    - slm['dfs'] (numpy.array) a (1,v) array containing the effective degrees of freedom.
+    - slm['resl'] (numpy.array) a (e,v) array containing the sum over observations of squares of differences of normalized residuals along each edge.
+    - slm['tri'] (numpy.array) a (v,3) array containing a mesh's triangle indices
+    - slm['lat'] (numpy.array) a 3D array of 1's and 0's where 1's are inside the lattice.
+
+    Note that slm['tri'] and slm['lat'] are mutually exclusive.
     """
     l, v = np.shape(slm['t'])
 
@@ -53,12 +62,12 @@ def _fdr(slm, mask=None):
         df[0, ndf-1] = slm['dfs'][0, mask > 0].mean()
 
     if 'du' in slm:
-        resels, reselspvert, edg = _resels(slm, mask.flatten())
+        resels, reselspvert, edg = resels(slm, mask.flatten())
     else:
         reselspvert = np.ones((v))
 
     varA = np.append(10, slm['t'][0, mask.astype(bool)])
-    P_val = _stat_threshold(df=df, p_val_peak=varA,
+    P_val = stat_threshold(df=df, p_val_peak=varA,
                            nvar=float(slm['k']), nprint=0)[0]
     P_val = P_val[1:len(P_val)]
     nx = len(P_val)
@@ -86,7 +95,7 @@ def _fdr(slm, mask=None):
     return qval
 
 
-def _random_field_theory(slm, mask=None, clusthresh=0.001):
+def random_field_theory(slm, mask=None, clusthresh=0.001):
     """Corrected P-values for vertices and clusters.
 
     Parameters
@@ -167,7 +176,7 @@ def _random_field_theory(slm, mask=None, clusthresh=0.001):
     if v == 1:
         varA = varA = np.concatenate((np.array([10]), slm['t'][0]))
         pval = {}
-        pval['P'] = _stat_threshold(df=df, p_val_peak=varA,
+        pval['P'] = stat_threshold(df=df, p_val_peak=varA,
                                    nvar=float(slm['k']), nprint=0)[0]
         pval['P'] = pval['P'][1]
         peak = []
@@ -177,7 +186,7 @@ def _random_field_theory(slm, mask=None, clusthresh=0.001):
         return pval, peak, clus, clusid
 
     if clusthresh < 1:
-        thresh = _stat_threshold(df=df, p_val_peak=clusthresh,
+        thresh = stat_threshold(df=df, p_val_peak=clusthresh,
                                 nvar=float(slm['k']), nprint=0)[0]
         thresh = float(thresh[0])
     else:
@@ -189,7 +198,7 @@ def _random_field_theory(slm, mask=None, clusthresh=0.001):
     if np.max(slm['t'][0, mask]) < thresh:
         pval = {}
         varA = np.concatenate((np.array([[10]]), slm['t']), axis=1)
-        pval['P'] = _stat_threshold(search_volume=resels, num_voxels=N,
+        pval['P'] = stat_threshold(search_volume=resels, num_voxels=N,
                                    fwhm=1, df=df,
                                    p_val_peak=varA.flatten(),
                                    nvar=float(slm['k']), nprint=0)[0]
@@ -198,12 +207,12 @@ def _random_field_theory(slm, mask=None, clusthresh=0.001):
         clus = []
         clusid = []
     else:
-        peak, clus, clusid = _peak_clus(slm, mask, thresh, reselspvert, edg)
+        peak, clus, clusid = peak_clus(slm, mask, thresh, reselspvert, edg)
         slm['t'] = slm['t'].reshape(1, slm['t'].size)
         varA = np.concatenate((np.array([[10]]), peak['t'].T, slm['t']),
                               axis=1)
         varB = np.concatenate((np.array([[10]]), clus['resels']))
-        pp, clpval, _, _, _, _, = _stat_threshold(search_volume=resels,
+        pp, clpval, _, _, _, _, = stat_threshold(search_volume=resels,
                                                  num_voxels=N, fwhm=1,
                                                  df=df,
                                                  p_val_peak=varA.flatten(),
@@ -224,7 +233,7 @@ def _random_field_theory(slm, mask=None, clusthresh=0.001):
                                      np.arange(0, slm['k'])/2)/ndf
             varA = np.convolve(resels.flatten(), sphere.flatten())
             varB = np.concatenate((np.array([[10]]), clus['resels']))
-            pp, clpval, _, _, _, _, = _stat_threshold(search_volume=varA,
+            pp, clpval, _, _, _, _, = stat_threshold(search_volume=varA,
                                                      num_voxels=math.inf, fwhm=1.0,
                                                      df=df, cluster_threshold=thresh,
                                                      p_val_extent=varB, nprint=0)
@@ -234,7 +243,7 @@ def _random_field_theory(slm, mask=None, clusthresh=0.001):
         y = np.concatenate((np.array([[1]]), clus['P']), axis=0)
         pval['C'] = interp1d(x.flatten(), y.flatten())(clusid)
 
-    tlim = _stat_threshold(search_volume=resels, num_voxels=N, fwhm=1,
+    tlim = stat_threshold(search_volume=resels, num_voxels=N, fwhm=1,
                           df=df, p_val_peak=np.array([0.5, 1]),
                           nvar=float(slm['k']), nprint=0)[0]
     tlim = tlim[1]
@@ -244,9 +253,9 @@ def _random_field_theory(slm, mask=None, clusthresh=0.001):
     return pval, peak, clus, clusid
 
 
-def _stat_threshold(search_volume=0, num_voxels=1, fwhm=0.0, df=math.inf,
-                    p_val_peak=0.05, cluster_threshold=0.001, p_val_extent=0.05, nconj=1,
-                    nvar=1, EC_file=None, expr=None, nprint=5):
+def stat_threshold(search_volume=0, num_voxels=1, fwhm=0.0, df=math.inf,
+                   p_val_peak=0.05, cluster_threshold=0.001, p_val_extent=0.05, nconj=1,
+                   nvar=1, EC_file=None, expr=None, nprint=5):
     """Thresholds and P-values of peaks and clusters of random fields in any D.
 
     Parameters
@@ -748,3 +757,568 @@ def _stat_threshold(search_volume=0, num_voxels=1, fwhm=0.0, df=math.inf,
                 print(P_val_extent_1)
 
     return peak_threshold, extent_threshold, peak_threshold_1, extent_threshold_1, t, rho
+
+
+def peak_clus(slm, mask, thresh, reselspvert=None, edg=None):
+    """ Finds peaks (local maxima) and clusters for surface data.
+    Parameters
+    ----------
+    slm : a dictionary, mandatory keys: 't', 'tri' (or 'lat'),
+        optional keys 'df', 'k'.
+        slm['t'] : numpy array of shape (l,v),
+            v is the number of vertices, the first row slm['t'][0,:] is used
+            for the clusters, and the other rows are used to calculate cluster
+            resels if slm['k']>1. See F for the precise definition
+            of the extra rows.
+        slm['tri'] : numpy array of shape (t,3), dype=int,
+            triangle indices, values should be 1 and v,
+        or,
+        slm['lat'] : numpy array of shape (nx,nx,nz),
+            values should be either 0 or 1.
+            note that [nx,ny,nz]=size(volume).
+        mask : numpy array of shape (v), dytpe=int,
+            values should be either 0 or 1.
+        thresh : float,
+            clusters are vertices where slm['t'][0,mask]>=thresh.
+        reselspvert : numpy array of shape (v),
+            resels per vertex, by default: np.ones(v).
+        edg :  numpy array of shape (e,2), dtype=int,
+            edge indices, by default computed from mesh_edges function.
+        slm['df'] : int,
+            degrees of freedom, note that only the length (1 or 2) is used
+            to determine if slm['t'] is Hotelling's T or T^2 when k>1.
+        slm['k'] : int,
+             k is number of variates, by default 1.
+
+    Returns
+    -------
+    peak : a dictionary with keys 't', 'vertid', 'clusid'.
+        peak['t'] : numpy array of shape (np,1),
+            array of peaks (local maxima).
+        peak['vertid] : numpy array of shape (np,1),
+            array of vertex id's (1-based).
+        peak['clusid'] : numpy array of shape (np,1),
+            array of cluster id's that contain the peak.
+    clus : a dictionary with keys 'clusid', 'nverts', 'resels'.
+        clus['clusid'] : numpy array of shape (nc,1),
+            array of cluster id numbers.
+        clus['nverts'] : numpy array of shape (nc,1),
+            array of number of vertices in the cluster.
+        clus['resels'] : numpy array of shape (nc,1),
+            array of resels in the cluster.
+    clusid : numpy array of shape (1,v),
+        array of cluster id's for each vertex.
+    """
+    if edg is None:
+        edg = mesh_edges(slm)
+
+    l, v = np.shape(slm['t'])
+    slm_t = copy.deepcopy(slm['t'])
+    slm_t[0, ~mask.astype(bool)] = slm_t[0,:].min()
+    t1 = slm_t[0, edg[:,0]]
+    t2 = slm_t[0, edg[:,1]]
+    islm = np.ones((1,v))
+    islm[0, edg[t1 < t2, 0]] = 0
+    islm[0, edg[t2 < t1, 1]] = 0
+    lmvox = np.argwhere(islm)[:,1] + 1
+    excurset = np.array(slm_t[0,:] >= thresh, dtype=int)
+    n = excurset.sum()
+
+    if n < 1:
+        peak = []
+        clus = []
+        clusid = []
+        return peak, clus, clusid
+
+    voxid = np.cumsum(excurset)
+    edg = voxid[edg[np.all(excurset[edg],1), :]]
+    nf = np.arange(1,n+1)
+
+    # Find cluster id's in nf (from Numerical Recipes in C, page 346):
+    for el in range(1, edg.shape[0]+1):
+        j = edg[el-1, 0]
+        k = edg[el-1, 1]
+        while nf[j-1] != j:
+            j = nf[j-1]
+        while nf[k-1] != k:
+            k = nf[k-1]
+        if j != k:
+            nf[j-1] = k
+
+    for j in range(1, n+1):
+        while nf[j-1] != nf[nf[j-1]-1]:
+            nf[j-1] =  nf[nf[j-1]-1]
+
+    vox = np.argwhere(excurset) + 1
+    ivox = np.argwhere(np.in1d(vox, lmvox)) + 1
+    clmid = nf[ivox-1]
+    uclmid, iclmid, jclmid = np.unique(clmid,
+                                       return_index=True, return_inverse=True)
+    iclmid = iclmid +1
+    jclmid = jclmid +1
+    ucid = np.unique(nf)
+    nclus = len(ucid)
+    # implementing matlab's histc function ###
+    bin_edges   = np.r_[-np.Inf, 0.5 * (ucid[:-1] + ucid[1:]), np.Inf]
+    ucvol, ucvol_edges = np.histogram(nf, bin_edges)
+
+    if reselspvert is None:
+        reselsvox = np.ones(np.shape(vox))
+    else:
+        reselsvox = reselspvert[vox-1]
+
+    # calling matlab-python version for scipy's interp1d
+    nf1 = interp1(np.append(0, ucid), np.arange(0,nclus+1), nf,
+                      kind='nearest')
+
+    # if k>1, find volume of cluster in added sphere
+    if 'k' not in slm or slm['k'] == 1:
+        ucrsl = np.bincount(nf1.astype(int), reselsvox.flatten())
+    if 'k' in slm and slm['k'] == 2:
+        if l == 1:
+            ndf = len(np.array([slm['df']]))
+            r = 2 * np.arccos((thresh / slm_t[0, vox-1])**(float(1)/ndf))
+        else:
+            r = 2 * np.arccos(np.sqrt((thresh - slm_t[1,vox-1]) *
+                                      (thresh >= slm_t[1,vox-1]) /
+                                      (slm_t[0,vox-1] - slm_t[1,vox-1])))
+        ucrsl =  np.bincount(nf1.astype(int), (r.T * reselsvox.T).flatten())
+    if 'k' in slm and slm['k'] == 3:
+        if l == 1:
+            ndf = len(np.array([slm['df']]))
+            r = 2 * math.pi * (1 - (thresh / slm_t[0, vox-1])**
+                                (float(1)/ndf))
+        else:
+            nt = 20
+            theta = (np.arange(1,nt+1,1) - 1/2) / nt * math.pi / 2
+            s = (np.cos(theta)**2 * slm_t[1, vox-1]).T
+            if l == 3:
+                s =  s + ((np.sin(theta)**2) * slm_t[2,vox-1]).T
+            r = 2 * math.pi * (1 - np.sqrt((thresh-s)*(thresh>=s) /
+                                           (np.ones((nt,1)) *
+                                            slm_t[0, vox-1].T -
+                                            s ))).mean(axis=0)
+        ucrsl = np.bincount(nf1.astype(int), (r.T * reselsvox.T).flatten())
+
+    # and their ranks (in ascending order)
+    iucrls = sorted(range(len(ucrsl[1:])), key=lambda k: ucrsl[1:][k])
+    rankrsl = np.zeros((1, nclus))
+    rankrsl[0, iucrls] =  np.arange(nclus,0,-1)
+
+    lmid = lmvox[ismember(lmvox, vox)[0]]
+
+    varA = slm_t[0, (lmid-1)]
+    varB = lmid
+    varC = rankrsl[0,jclmid-1]
+    varALL = np.concatenate((varA.reshape(len(varA),1),
+                             varB.reshape(len(varB),1),
+                             varC.reshape(len(varC),1)), axis=1)
+    lm = np.flipud(varALL[varALL[:,0].argsort(),])
+    varNEW = np.concatenate((rankrsl.T, ucvol.reshape(len(ucvol),1),
+                             ucrsl.reshape(len(ucrsl),1)[1:]) , axis=1)
+    cl = varNEW[varNEW[:,0].argsort(),]
+    clusid = np.zeros((1,v))
+    clusid[0,(vox-1).T] = interp1(np.append(0, ucid),
+                                      np.append(0, rankrsl), nf,
+                                      kind='nearest')
+    peak = {}
+    peak['t'] = lm[:,0].reshape(len(lm[:,0]), 1)
+    peak['vertid'] = lm[:,1].reshape(len(lm[:,1]), 1)
+    peak['clusid'] = lm[:,2].reshape(len(lm[:,2]), 1)
+    clus = {}
+    clus['clusid'] = cl[:,0].reshape(len(cl[:,0]), 1)
+    clus['nverts'] = cl[:,1].reshape(len(cl[:,1]), 1)
+    clus['resels'] = cl[:,2] .reshape(len(cl[:,2]), 1)
+
+    return peak, clus, clusid
+
+
+def resels(slm, mask=None):
+    """ SurfStatResels of surface or volume data inside a mask.
+
+    Parameters
+    ----------
+    slm : a dictionary with keys 'lat' or 'tri' and, optionally, 'resl'.
+        slm['lat'] : a 3D numpy array of 1's and 0's.
+        slm['tri'] : a 2D numpy array of shape (t, 3).
+            Contains triangles of a surface. slm['tri'].max() is equal to the
+            number of vertices.
+        slm['resl'] : a 2D numpy array of shape (e, k).
+            Sum over observations of squares of differences of normalized
+            residuals along each edge.
+    mask : a 1D numpy array of shape (v), dtype 'bool'.
+        v must be equal to int(slm['tri'].max()).
+        Contains 1's and 0's (1's are included and 0's are excluded).
+
+    Returns
+    -------
+    resels : a 2D numpy array of shape (1, (D+1)).
+        Array of 0,...,D dimensional resels of the mask, EC of the mask
+        if slm['resl'] is not given.
+    reselspvert : a 1D numpy array of shape (v).
+        Array of D-dimensional resels per mask vertex.
+    edg : a 2D numpy array of shape (e, 2).
+        Array of edge indices.
+    """
+
+    def pacos(x):
+        return np.arccos( np.minimum(np.abs(x),1) * np.sign(x) )
+
+    if 'tri' in slm:
+        # Get unique edges. Subtract 1 from edges to conform to Python's
+        # counting from 0 - RV
+        tri = np.sort(slm['tri'])-1
+        edg = np.unique(np.vstack((tri[:,(0,1)], tri[:,(0,2)],
+                                       tri[:,(1,2)])),axis=0)
+
+
+        # If no mask is provided, create one with all included vertices set to
+        # 1. If mask is provided, simply grab the number of vertices from mask.
+        if mask is None:
+            v = np.amax(edg)+1
+            mask = np.full(v,False)
+            mask[edg-1] = True
+        else:
+            #if np.ndim(mask) > 1:
+            #    mask = np.squeeze(mask)
+            #    if mask.shape[0] > 1:
+            #        mask = mask.T
+            v = mask.size
+
+        ## Compute the Lipschitz–Killing curvatures (LKC)
+        m = np.sum(mask)
+        if 'resl' in slm:
+            lkc = np.zeros((3,3))
+        else:
+            lkc = np.zeros((1,3))
+        lkc[0,0] = m
+
+        # LKC of edges
+        maskedg = np.all(mask[edg],axis=1)
+        lkc[0,1] = np.sum(maskedg)
+
+        if 'resl' in slm:
+            r1 = np.mean(np.sqrt(slm['resl'][maskedg,:]),axis=1)
+            lkc[1,1] = np.sum(r1)
+        # LKC of triangles
+        # Made an adjustment from the MATLAB implementation:
+        # The reselspvert computation is included in the if-statement.
+        # MATLAB errors when the if statement is false as variable r2 is not
+        # defined during the computation of reselspvert. - RV
+        masktri = np.all(mask[tri],1)
+        lkc[0,2] = np.sum(masktri)
+        if 'resl' in slm:
+            loc = row_ismember(tri[masktri,:][:,[0,1]], edg)
+            l12 = slm['resl'][loc,:]
+            loc = row_ismember(tri[masktri,:][:,[0,2]], edg)
+            l13 = slm['resl'][loc,:]
+            loc = row_ismember(tri[masktri,:][:,[1,2]], edg)
+            l23 = slm['resl'][loc,:]
+            a = np.fmax(4*l12*l13-(l12+l13-l23)**2,0)
+            r2 = np.mean(np.sqrt(a),axis=1)/4
+            lkc[1,2] = np.sum(np.mean(np.sqrt(l12) +
+               np.sqrt(l13)+np.sqrt(l23),axis=1))/2
+            lkc[2,2] = np.nansum(r2,axis=0)
+
+            # Compute resels per mask vertex
+            reselspvert = np.zeros(v)
+            for j in range(0,3):
+                reselspvert = reselspvert + \
+                        np.bincount(tri[masktri,j], weights=r2, minlength=v)
+            D = 2
+            reselspvert = reselspvert / (D+1) / np.sqrt(4*np.log(2)) ** D
+        else:
+            reselspvert = None
+
+    if 'lat' in slm:
+        edg = mesh_edges(slm)
+        # The lattice is filled with 5 alternating tetrahedra per cube
+        I, J, K = np.shape(slm['lat'])
+        IJ = I*J
+        i, j = np.meshgrid(range(1,I+1),range(1,J+1))
+        i = np.squeeze(np.reshape(i,(-1,1)))
+        j = np.squeeze(np.reshape(j,(-1,1)))
+
+        c1  = np.argwhere((((i+j)%2)==0) & (i < I) & (j < J))
+        c2  = np.argwhere((((i+j)%2)==0) & (i > 1) & (j < J))
+        c11 = np.argwhere((((i+j)%2)==0) & (i == I) & (j < J))
+        c21 = np.argwhere((((i+j)%2)==0) & (i == I) & (j > 1))
+        c12 = np.argwhere((((i+j)%2)==0) & (i < I) & (j == J))
+        c22 = np.argwhere((((i+j)%2)==0) & (i > 1) & (j == J))
+
+        # outcome is 1 lower than MATLAB due to 0-1 counting difference. - RV
+        d1  = np.argwhere((((i+j)%2)==1) & (i < I) & (j < J))+IJ
+        d2  = np.argwhere((((i+j)%2)==1) & (i > 1) & (j < J))+IJ
+
+        tri1 = cat((
+            cat((c1, c1+1, c1+1+I),axis=1),
+            cat((c1, c1+I, c1+1+I),axis=1),
+            cat((c2-1, c2, c2-1+I),axis=1),
+            cat((c2, c2-1+I, c2+I),axis=1)),
+            axis=0)
+        tri2= cat((
+            cat((c1,    c1+1,    c1+1+IJ),axis=1),
+            cat((c1,    c1+IJ,   c1+1+IJ),axis=1),
+            cat((c1,    c1+I,    c1+I+IJ),axis=1),
+            cat((c1,     c1+IJ,   c1+I+IJ),axis=1),
+            cat((c1,     c1+1+I,  c1+1+IJ),axis=1),
+            cat((c1,     c1+1+I,  c1+I+IJ),axis=1),
+            cat((c1,     c1+1+IJ, c1+I+IJ),axis=1),
+            cat((c1+1+I, c1+1+IJ, c1+I+IJ),axis=1),
+            cat((c2-1,   c2,      c2-1+IJ),axis=1),
+            cat((c2,     c2-1+IJ, c2+IJ),axis=1),
+            cat((c2-1,   c2-1+I,  c2-1+IJ),axis=1),
+            cat((c2-1+I, c2-1+IJ, c2-1+I+IJ),axis=1),
+            cat((c2,     c2-1+I,  c2+I+IJ),axis=1),
+            cat((c2,     c2-1+IJ, c2+I+IJ),axis=1),
+            cat((c2,     c2-1+I,  c2-1+IJ),axis=1),
+            cat((c2-1+I, c2-1+IJ, c2+I+IJ),axis=1),
+            cat((c11,    c11+I,    c11+I+IJ),axis=1),
+            cat((c11,    c11+IJ,   c11+I+IJ),axis=1),
+            cat((c21-I,  c21,      c21-I+IJ),axis=1),
+            cat((c21,    c21-I+IJ, c21+IJ),axis=1),
+            cat((c12,    c12+1,    c12+1+IJ),axis=1),
+            cat((c12,    c12+IJ,   c12+1+IJ),axis=1),
+            cat((c22-1,  c22,      c22-1+IJ),axis=1),
+            cat((c22,    c22-1+IJ, c22+IJ),axis=1)),
+            axis=0)
+        tri3 = cat((
+            cat((d1,     d1+1,    d1+1+I),axis=1),
+            cat((d1,     d1+I,    d1+1+I),axis=1),
+            cat((d2-1,   d2,      d2-1+I),axis=1),
+            cat((d2,     d2-1+I,  d2+I),axis=1)),
+            axis=0)
+        tet1 = cat((
+            cat((c1,     c1+1,    c1+1+I,    c1+1+IJ),axis=1),
+            cat((c1,     c1+I,    c1+1+I,    c1+I+IJ),axis=1),
+            cat((c1,     c1+1+I,  c1+1+IJ,   c1+I+IJ),axis=1),
+            cat((c1,     c1+IJ,   c1+1+IJ,   c1+I+IJ),axis=1),
+            cat((c1+1+I, c1+1+IJ, c1+I+IJ,   c1+1+I+IJ),axis=1),
+            cat((c2-1,   c2,      c2-1+I,    c2-1+IJ),axis=1),
+            cat((c2,     c2-1+I,  c2+I,      c2+I+IJ),axis=1),
+            cat((c2,     c2-1+I,  c2-1+IJ,   c2+I+IJ),axis=1),
+            cat((c2,     c2-1+IJ, c2+IJ,     c2+I+IJ),axis=1),
+            cat((c2-1+I, c2-1+IJ, c2-1+I+IJ, c2+I+IJ),axis=1)),
+            axis=0)
+
+        v = np.int(np.round(np.sum(slm['lat'])))
+        if mask is None:
+            mask = np.ones(v,dtype=bool)
+
+        reselspvert = np.zeros(v)
+        vs = np.cumsum(np.squeeze(np.sum(np.sum(slm['lat'],axis=0),axis=0)))
+        vs = cat((np.zeros(1),vs,np.expand_dims(vs[K-1],axis=0)),axis=0)
+        vs = vs.astype(int)
+        es = 0
+        lat = np.zeros((I,J,2))
+        lat[:,:,0] = slm['lat'][:,:,0]
+        lkc = np.zeros((4,4))
+        for k in range(0,K):
+            f = (k+1) % 2
+            if k < (K-1):
+                lat[:,:,f] = slm['lat'][:,:,k+1]
+            else:
+                lat[:,:,f] = np.zeros((I,J))
+            vid = (np.cumsum(lat.flatten('F')) * np.reshape(lat.T,-1)).astype(int)
+            if f:
+                edg1 = edg[np.logical_and(edg[:,0]>(vs[k]-1), \
+                                          edg[:,0] <= (vs[k+1]-1)),:]-vs[k]
+                edg2 = edg[np.logical_and(edg[:,0] > (vs[k]-1), \
+                                          edg[:,1] <= (vs[k+2]-1)),:]-vs[k]
+                # Added a -1 - RV
+                tri = cat((vid[tri1[np.all(np.reshape(lat.flatten('F')[tri1], \
+                                                      tri1.shape),1),:]],
+                           vid[tri2[np.all(np.reshape(lat.flatten('F')[tri2], \
+                                                      tri2.shape),1),:]]),
+                           axis=0)-1
+                mask1 = mask[np.arange(vs[k],vs[k+2])]
+            else:
+                edg1 = cat((
+                    edg[np.logical_and(edg[:,0]  > (vs[k]-1), edg[:,1] <= \
+                        (vs[k+1]-1)), :] - vs[k] + vs[k+2] - vs[k+1],
+                    cat((
+                        np.expand_dims(edg[np.logical_and(edg[:,0] <=        \
+                                                          (vs[k+1]-1),       \
+                                                          edg[:,1] >         \
+                                                          (vs[k+1]-1)), 1]   \
+                                                          - vs[k+1],axis=1),
+                        np.expand_dims(edg[np.logical_and(edg[:,0] <=        \
+                                                          (vs[k+1]-1),       \
+                                                          edg[:,1] >         \
+                                                          (vs[k+1]-1)), 0]   \
+                                                          - vs[k] + vs[k+2]  \
+                                                          - vs[k+1],axis=1)),
+                        axis=1)),
+                    axis=0)
+                edg2 = cat((edg1, edg[np.logical_and(edg[:,0] > (vs[k+1]-1), \
+                           edg[:,1] <= (vs[k+2]-1)),:] - vs[k+1]), axis=0)
+                # Added a -1 - RV
+                tri = cat((vid[tri3[np.all(lat.flatten('F')[tri3],axis=1),:]],
+                          vid[tri2[np.all(lat.flatten('F')[tri2],axis=1),:]]),
+                          axis=0)-1
+                mask1 = cat((
+                    mask[np.arange(vs[k+1], vs[k+2])],
+                    mask[np.arange(vs[k],   vs[k+1])]))
+            # Added a -1 -RV
+            tet = vid[tet1[np.all(lat.flatten('F')[tet1],axis=1),:]]-1
+            m1 = np.max(edg2[:,0])
+            ue = edg2[:,0] + m1 * (edg2[:,1]-1)
+            e = edg2.shape[0]
+            ae = np.arange(0,e)
+            if e < 2 ** 31:
+                sparsedg = csr_matrix((ae,(ue,np.zeros(ue.shape,dtype=int))),
+                                      dtype=np.int)
+                sparsedg.eliminate_zeros()
+            ##
+            lkc1 = np.zeros((4,4))
+            lkc1[0,0] = np.sum(mask[np.arange(vs[k],vs[k+1])])
+
+            ## LKC of edges
+            maskedg = np.all(mask1[edg1],axis=1)
+
+            lkc1[0,1] = np.sum(maskedg)
+            if 'resl' in slm:
+                r1 = np.mean(np.sqrt(slm['resl'][np.argwhere(maskedg)+es,:]),
+                             axis=1)
+                lkc1[1,1] = np.sum(r1)
+
+            ## LKC of triangles
+            masktri = np.all(mask1[tri],axis=1).flatten()
+            lkc1[0,2] = np.sum(masktri)
+            if 'resl' in slm:
+                if all(masktri == False):
+                    # Set these variables to empty arrays to match the MATLAB
+                    # implementation.
+                    lkc1[1,2] = 0
+                    lkc1[2,2] = 0
+                else:
+                    if e < 2 ** 31:
+                        l12 = slm['resl'][sparsedg[tri[masktri,0] + m1 * \
+                                        (tri[masktri,1]-1), 0].toarray() + es, :]
+                        l13 = slm['resl'][sparsedg[tri[masktri,0] + m1 * \
+                                        (tri[masktri,2]-1), 0].toarray() + es, :]
+                        l23 = slm['resl'][sparsedg[tri[masktri,1] + m1 * \
+                                        (tri[masktri,2]-1), 0].toarray() + es, :]
+                    else:
+                        l12 = slm['resl'][interp1(ue,ae,tri[masktri,0] + m1 * \
+                                (tri[masktri,1] - 1),kind='nearest') + es, :]
+                        l13 = slm['resl'][interp1(ue,ae,tri[masktri,0] + m1 * \
+                                (tri[masktri,2] - 1),kind='nearest') + es, :]
+                        l23 = slm['resl'][interp1(ue,ae,tri[masktri,1] + m1 * \
+                                (tri[masktri,2] - 1),kind='nearest') + es, :]
+                    a = np.fmax(4 * l12 * l13 - (l12+l13-l23) ** 2, 0)
+                    r2 = np.mean(np.sqrt(a),axis=1)/4
+                    lkc1[1,2] = np.sum(np.mean(np.sqrt(l12) + np.sqrt(l13) +
+                                    np.sqrt(l23),axis=1))/2
+                    lkc1[2,2] = np.sum(r2)
+
+                # The following if-statement has nargout >=2 in MATLAB,
+                # but there's no Python equivalent so ignore that. - RV
+                if K == 1:
+                    for j in range(0,3):
+                        if f:
+                            v1 = tri[masktri,j] + vs[k]
+                        else:
+                            v1 = tri[masktri,j] + vs[k+1]
+                            v1 = v1 - int(vs > vs[k+2]) * (vs[k+2]-vs[k])
+                        reselspvert += np.bincount(v1, r2, v)
+
+            ## LKC of tetrahedra
+            masktet = np.all(mask1[tet],axis=1).flatten()
+            lkc1[0,3] = np.sum(masktet)
+            if 'resl' in slm and k < (K-1):
+                if e < 2 ** 31:
+                    l12 = slm['resl'][(sparsedg[tet[masktet,0] + m1 * \
+                                  (tet[masktet,1]-1),0].toarray() + es).tolist(), :]
+                    l13 = slm['resl'][(sparsedg[tet[masktet,0] + m1 * \
+                                  (tet[masktet,2]-1),0].toarray() + es).tolist(), :]
+                    l23 = slm['resl'][(sparsedg[tet[masktet,1] + m1 * \
+                                  (tet[masktet,2]-1),0].toarray() + es).tolist(), :]
+                    l14 = slm['resl'][(sparsedg[tet[masktet,0] + m1 * \
+                                  (tet[masktet,3]-1),0].toarray() + es).tolist(), :]
+                    l24 = slm['resl'][(sparsedg[tet[masktet,1] + m1 * \
+                                  (tet[masktet,3]-1),0].toarray() + es).tolist(), :]
+                    l34 = slm['resl'][(sparsedg[tet[masktet,2] + m1 * \
+                                  (tet[masktet,3]-1),0].toarray() + es).tolist(), :]
+                else:
+                    l12 = slm['resl'][interp1(ue,ae,tet[masktet,0] + m1 * \
+                              (tet[masktet,1]-1),kind='nearest')+es,:]
+                    l13 = slm['resl'][interp1(ue,ae,tet[masktet,0] + m1 * \
+                              (tet[masktet,2]-1),kind='nearest')+es,:]
+                    l23 = slm['resl'][interp1(ue,ae,tet[masktet,1] + m1 * \
+                              (tet[masktet,2]-1),kind='nearest')+es,:]
+                    l14 = slm['resl'][interp1(ue,ae,tet[masktet,0] + m1 * \
+                              (tet[masktet,3]-1),kind='nearest')+es,:]
+                    l24 = slm['resl'][interp1(ue,ae,tet[masktet,1] + m1 * \
+                              (tet[masktet,3]-1),kind='nearest')+es,:]
+                    l34 = slm['resl'][interp1(ue,ae,tet[masktet,2] + m1 * \
+                              (tet[masktet,3]-1),kind='nearest')+es,:]
+                a4 = np.fmax(4 * l12 * l13 - (l12 + l13 -l23) ** 2, 0)
+                a3 = np.fmax(4 * l12 * l14 - (l12 + l14 -l24) ** 2, 0)
+                a2 = np.fmax(4 * l13 * l14 - (l13 + l14 -l34) ** 2, 0)
+                a1 = np.fmax(4 * l23 * l24 - (l23 + l24 -l34) ** 2, 0)
+
+                d12 = 4 * l12 * l34 - (l13 + l24 - l23 - l14) ** 2
+                d13 = 4 * l13 * l24 - (l12 + l34 - l23 - l14) ** 2
+                d14 = 4 * l14 * l23 - (l12 + l34 - l24 - l13) ** 2
+
+                h = np.logical_or(a1 <= 0, a2 <= 0)
+                delta12 = np.sum(np.mean(np.sqrt(l34) * pacos((d12-a1-a2) / \
+                            np.sqrt(a1 * a2 + h) / 2 * (1-h) + h),axis=1))
+                h = np.logical_or(a1 <= 0, a3 <= 0)
+                delta13 = np.sum(np.mean(np.sqrt(l24) * pacos((d13-a1-a3) / \
+                            np.sqrt(a1 * a3 + h) / 2 * (1-h) + h),axis=1))
+                h = np.logical_or(a1 <= 0, a4 <= 0)
+                delta14 = np.sum(np.mean(np.sqrt(l23) * pacos((d14-a1-a4) / \
+                            np.sqrt(a1 * a4 + h) / 2 * (1-h) + h),axis=1))
+                h = np.logical_or(a2 <= 0, a3 <= 0)
+                delta23 = np.sum(np.mean(np.sqrt(l14) * pacos((d14-a2-a3) / \
+                            np.sqrt(a2 * a3 + h) / 2 * (1-h) + h),axis=1))
+                h = np.logical_or(a2 <= 0, a4 <= 0)
+                delta24 = np.sum(np.mean(np.sqrt(l13) * pacos((d13-a2-a4) / \
+                            np.sqrt(a2 * a4 + h) / 2 * (1-h) + h),axis=1))
+                h = np.logical_or(a3 <= 0, a4 <= 0)
+                delta34 = np.sum(np.mean(np.sqrt(l12) * pacos((d12-a3-a4) / \
+                            np.sqrt(a3 * a4 + h) / 2 * (1-h) + h),axis=1))
+
+                r3 = np.squeeze(np.mean(np.sqrt(np.fmax((4 * a1 * a2 - \
+                                (a1 + a2 - d12) **2) / (l34 + (l34<=0)) * \
+                                (l34>0), 0)),axis=1) / 48)
+
+                lkc1[1,3] = (delta12 + delta13 + delta14 + delta23 + delta24 +
+                            delta34)/(2 * np.pi)
+                lkc1[2,3] = np.sum(np.mean(np.sqrt(a1) + np.sqrt(a2) +
+                            np.sqrt(a3) + np.sqrt(a4), axis=1))/8
+                lkc1[3,3] = np.sum(r3)
+
+                ## Original MATLAB code has a if nargout>=2 here, ignore it
+                # as no equivalent exists in Python - RV.
+                for j in range(0,4):
+                    if f:
+                        v1 = tet[masktet,j] + vs[k]
+                    else:
+                        v1 = tet[masktet,j] + vs[k+1]
+                        v1 = v1 - (v1 > (vs[k+2]-1)) * (vs[k+2] - vs[k])
+                    if np.ndim(r3) == 0:
+                        r3 = r3.tolist()
+                        r3 = [r3]
+                    reselspvert += np.bincount(v1, r3, v)
+            lkc = lkc + lkc1
+            es = es + edg1.shape[0]
+
+        ## Original MATLAB code has a if nargout>=2 here,
+        # ignore it as no equivalent exists in Python - RV.
+        D = 2 + (K>1)
+        reselspvert = reselspvert / (D+1) / np.sqrt(4*np.log(2)) ** D
+
+    ## Compute resels - RV
+    D1 = lkc.shape[0]-1
+    D2 = lkc.shape[1]-1
+    tpltz = toeplitz((-1)**(np.arange(0,D1+1)), (-1)**(np.arange(0,D2+1)))
+    lkcs = np.sum(tpltz * lkc, axis=1).T
+    lkcs = np.trim_zeros(lkcs,trim='b')
+    lkcs = np.atleast_2d(lkcs)
+    D = lkcs.shape[1]-1
+    resels = lkcs / np.sqrt(4*np.log(2))**np.arange(0,D+1)
+
+    return resels, reselspvert, edg
